requirements.txt
aliyun-python-sdk-core==2.13.24
aliyun-python-sdk-dysmsapi==2.1.1
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==2.1.7
django-shortuuidfield==0.1.3
idna==2.10
jmespath==0.10.0
Pillow==7.2.0
pycryptodome==3.9.8
PyJWT==1.7.1
python-memcached==1.59
pytz==2020.1
requests==2.24.0
shortuuid==1.0.1
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
